REFACTOR : getMimeTypeAndExtension 모듈화
