+++ domain/src/main/java/com/fakedevelopers/domain/model/MediaInfo.kt
@@ -1,0 +1,6 @@
+package com.fakedevelopers.domain.model
+
+data class MediaInfo(
+    val mimeType: String,
+    val extension: String
+)
