+++ presentation/src/main/java/com/fakedevelopers/presentation/ui/util/BitmapUtil.kt
@@ -1,0 +1,13 @@
+package com.fakedevelopers.presentation.ui.util
+
+import android.graphics.Bitmap
+import android.graphics.BitmapFactory
+import android.graphics.Matrix
+
+const val ROTATE_DEGREE = 90f
+
+fun ByteArray.toBitmap(): Bitmap = BitmapFactory.decodeByteArray(this, 0, size)
+fun Bitmap.getRotated(degree: Float): Bitmap {
+    val matrix = Matrix().apply { postRotate(degree) }
+    return Bitmap.createBitmap(this, 0, 0, width, height, matrix, true) ?: this
+}
