+++ presentation/src/main/java/com/fakedevelopers/presentation/ui/util/UtilModule.kt
@@ -11,12 +11,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 object UtilModule {
-
-    @Singleton
-    @Provides
-    fun provideAlbumImageUtil(@ApplicationContext context: Context) =
-        AlbumImageUtils(context.contentResolver)
-
     @Singleton
     @Provides
     fun provideDateUtil(@ApplicationContext context: Context) = DateUtil(context)
